refactor: 위도 경도 double으로 수정
- 카카오 지도 API 기준 소수점이 13자리까지 존재함
- 따라서 float 타입에서 double 타입으로 변경

+++ src/main/java/com/elice/ustory/domain/paper/dto/UpdatePaperRequest.java
@@ -43,10 +43,10 @@ public class UpdatePaperRequest {
     private String store;
 
     @Schema(description = "X좌표", example = "37.5494")
-    private float coordinateX;
+    private double coordinateX;
 
     @Schema(description = "Y좌표", example = "126.9169")
-    private float coordinateY;
+    private double coordinateY;
 
     public Paper toPageEntity() {
         return Paper.createBuilder()
